Add Bulk of Sheets V4 API Functions
Add all functions that could be defined based on the Sheets V4 discovery document:  https://sheets.googleapis.com/$discovery/rest?version=v4. These functions are almost a verbatim translation of the documentation and arguments of the API itself so any functionality that the API exposes can be implemented through these functions. Although, these functions are sometimes verbose and complex so that wrappers will be written to reduce barriers to entry and align a common interface that googlesheets already implements.